Release 5.8. Production release

+++ source/backend/api/Services/AcquisitionFileService.cs
@@ -39,6 +39,7 @@ namespace Pims.Api.Services
         private readonly IAcquisitionStatusSolver _statusSolver;
         private readonly IPropertyService _propertyService;
         private readonly IProjectRepository _projectRepository;
+        private readonly IPropertyOperationService _propertyOperationService;
 
         public AcquisitionFileService(
             ClaimsPrincipal user,
@@ -58,7 +59,8 @@ namespace Pims.Api.Services
             ITakeRepository takeRepository,
             IProjectRepository projectRepository,
             IAcquisitionStatusSolver statusSolver,
-            IPropertyService propertyService)
+            IPropertyService propertyService,
+            IPropertyOperationService propertyOperationService)
         {
             _user = user;
             _logger = logger;
@@ -78,6 +80,7 @@ namespace Pims.Api.Services
             _statusSolver = statusSolver;
             _propertyService = propertyService;
             _projectRepository = projectRepository;
+            _propertyOperationService = propertyOperationService;
         }
 
         public Paged<PimsAcquisitionFile> GetPage(AcquisitionFilter filter)
@@ -260,9 +263,14 @@ namespace Pims.Api.Services
                 ValidateDraftsOnComplete(acquisitionFile);
             }
 
+            if(currentAcquisitionStatus != AcquisitionStatusTypes.CANCEL && acquisitionFile.AcquisitionFileStatusTypeCode == AcquisitionStatusTypes.CANCEL.ToString())
+            {
+                ValidateDraftsOnCancelled(acquisitionFile);
+            }
+
             if (!_statusSolver.CanEditDetails(currentAcquisitionStatus) && !_user.HasPermission(Permissions.SystemAdmin))
             {
-                throw new BusinessRuleViolationException("The file you are editing is not active or hold, so you cannot save changes. Refresh your browser to see file state.");
+                throw new BusinessRuleViolationException("The file you are editing is not active, so you cannot save changes. Refresh your browser to see file state.");
             }
 
             if (!userOverrides.Contains(UserOverrideCode.UpdateRegion))
@@ -311,9 +319,9 @@ namespace Pims.Api.Services
             ValidatePropertyRegions(acquisitionFile);
 
             AcquisitionStatusTypes? currentAcquisitionStatus = GetCurrentAcquisitionStatus(acquisitionFile.Internal_Id);
-            if (!_statusSolver.CanEditProperties(currentAcquisitionStatus) && !_user.HasPermission(Permissions.SystemAdmin))
+            if (!_statusSolver.CanEditProperties(currentAcquisitionStatus))
             {
-                throw new BusinessRuleViolationException("The file you are editing is not active or hold, so you cannot save changes. Refresh your browser to see file state.");
+                throw new BusinessRuleViolationException("The file you are editing is not active, so you cannot save changes. Refresh your browser to see file state.");
             }
 
             // Get the current properties in the acquisition file
@@ -327,6 +335,7 @@ namespace Pims.Api.Services
                 {
                     incomingAcquisitionProperty.Internal_Id = matchingProperty.Internal_Id;
                 }
+
                 // If the property is not new, check if the name has been updated.
                 if (incomingAcquisitionProperty.Internal_Id != 0)
                 {
@@ -364,7 +373,7 @@ namespace Pims.Api.Services
             foreach (var deletedProperty in differenceSet)
             {
                 var acqFileProperties = _acquisitionFilePropertyRepository.GetPropertiesByAcquisitionFileId(acquisitionFile.Internal_Id).FirstOrDefault(ap => ap.PropertyId == deletedProperty.PropertyId);
-                if (acqFileProperties.PimsTakes.Any() || acqFileProperties.PimsInthldrPropInterests.Any())
+                if (acqFileProperties.PimsTakes.Count > 0 || acqFileProperties.PimsInthldrPropInterests.Count > 0)
                 {
                     throw new BusinessRuleViolationException("You must remove all takes and interest holders from an acquisition file property before removing that property from an acquisition file");
                 }
@@ -372,6 +381,11 @@ namespace Pims.Api.Services
                 if (_acquisitionFilePropertyRepository.AcquisitionFilePropertyInCompensationReq(deletedProperty.PropertyAcquisitionFileId))
                 {
                     throw new BusinessRuleViolationException("Acquisition File property can not be removed since it's assigned as a property for a compensation requisition");
+                }
+
+                if (_propertyOperationService.GetOperationsForProperty(deletedProperty.PropertyId).Count > 0)
+                {
+                    throw new BusinessRuleViolationException("This property cannot be deleted because it is part of a subdivision or consolidation");
                 }
 
                 _acquisitionFilePropertyRepository.Delete(deletedProperty);
@@ -403,9 +417,9 @@ namespace Pims.Api.Services
             _user.ThrowInvalidAccessToAcquisitionFile(_userRepository, _acqFileRepository, acquisitionFileId);
 
             var currentAcquisitionStatus = GetCurrentAcquisitionStatus(acquisitionFileId);
-            if (!_statusSolver.CanEditChecklists(currentAcquisitionStatus) && !_user.HasPermission(Permissions.SystemAdmin))
+            if (!_statusSolver.CanEditChecklists(currentAcquisitionStatus))
             {
-                throw new BusinessRuleViolationException("The file you are editing is not active or hold, so you cannot save changes. Refresh your browser to see file state.");
+                throw new BusinessRuleViolationException("The file you are editing is not active, so you cannot save changes. Refresh your browser to see file state.");
             }
 
             // Get the current checklist items for this acquisition file.
@@ -447,7 +461,7 @@ namespace Pims.Api.Services
             _user.ThrowIfNotAuthorized(Permissions.AgreementView);
             _user.ThrowInvalidAccessToAcquisitionFile(_userRepository, _acqFileRepository, acquisitionFileId);
 
-            ValidateAcquisitionFileStatus(acquisitionFileId, agreement.AgreementStatusTypeCode);
+            ValidateAcquisitionFileStatusForAgreement(acquisitionFileId);
 
             var newAgreement = _agreementRepository.AddAgreement(agreement);
             _agreementRepository.CommitTransaction();
@@ -483,9 +497,7 @@ namespace Pims.Api.Services
             _user.ThrowIfNotAuthorized(Permissions.AgreementView);
             _user.ThrowInvalidAccessToAcquisitionFile(_userRepository, _acqFileRepository, acquisitionFileId);
 
-            var currentAgreement = _agreementRepository.GetAgreementById(agreement.AgreementId);
-
-            ValidateAcquisitionFileStatus(acquisitionFileId, currentAgreement.AgreementStatusTypeCode);
+            ValidateAcquisitionFileStatusForAgreement(acquisitionFileId);
 
             var updatedAgreement = _agreementRepository.UpdateAgreement(agreement);
             _agreementRepository.CommitTransaction();
@@ -498,10 +510,7 @@ namespace Pims.Api.Services
             _user.ThrowIfNotAuthorized(Permissions.AgreementView);
             _user.ThrowInvalidAccessToAcquisitionFile(_userRepository, _acqFileRepository, acquisitionFileId);
 
-            var currentAcquisitionStatus = GetCurrentAcquisitionStatus(acquisitionFileId);
-            var agreement = _agreementRepository.GetAgreementById(agreementId);
-
-            ValidateAcquisitionFileStatus(acquisitionFileId, agreement.AgreementStatusTypeCode);
+            ValidateAcquisitionFileStatusForAgreement(acquisitionFileId);
 
             bool deleteResult = _agreementRepository.TryDeleteAgreement(acquisitionFileId, agreementId);
             _agreementRepository.CommitTransaction();
@@ -525,9 +534,9 @@ namespace Pims.Api.Services
             _user.ThrowInvalidAccessToAcquisitionFile(_userRepository, _acqFileRepository, acquisitionFileId);
 
             var currentAcquisitionStatus = GetCurrentAcquisitionStatus(acquisitionFileId);
-            if (!_statusSolver.CanEditStakeholders(currentAcquisitionStatus) && !_user.HasPermission(Permissions.SystemAdmin))
+            if (!_statusSolver.CanEditStakeholders(currentAcquisitionStatus))
             {
-                throw new BusinessRuleViolationException("The file you are editing is not active or hold, so you cannot save changes. Refresh your browser to see file state.");
+                throw new BusinessRuleViolationException("The file you are editing is not active, so you cannot save changes. Refresh your browser to see file state.");
             }
 
             var currentInterestHolders = _interestHolderRepository.GetInterestHoldersByAcquisitionFile(acquisitionFileId);
@@ -573,6 +582,12 @@ namespace Pims.Api.Services
             if (acquisitionFileId != expPayment.AcquisitionFileId || acquisitionFileParent is null)
             {
                 throw new BadRequestException("Invalid acquisitionFileId.");
+            }
+
+            var currentAcquisitionStatus = GetCurrentAcquisitionStatus(acquisitionFileId);
+            if (!_statusSolver.CanEditExpropriation(currentAcquisitionStatus))
+            {
+                throw new BusinessRuleViolationException("The file you are editing is not active, so you cannot save changes. Refresh your browser to see file state.");
             }
 
             var newForm8 = _expropriationPaymentRepository.Add(expPayment);
@@ -748,14 +763,23 @@ namespace Pims.Api.Services
             }
         }
 
-        private void ValidateAcquisitionFileStatus(long acquisitionFileId, string agreementCurrentStatusTypeCode)
+        private void ValidateDraftsOnCancelled(PimsAcquisitionFile incomingFile)
+        {
+            var takes = _takeRepository.GetAllByAcquisitionFileId(incomingFile.AcquisitionFileId);
+
+            if (takes.Any(t => t.TakeStatusTypeCode == AcquisitionTakeStatusTypes.INPROGRESS.ToString()))
+            {
+                throw new BusinessRuleViolationException("Please ensure all in-progress property takes have been completed or cancelled before cancelling an Acquisition File.");
+            }
+        }
+
+        private void ValidateAcquisitionFileStatusForAgreement(long acquisitionFileId)
         {
             var currentAcquisitionStatus = GetCurrentAcquisitionStatus(acquisitionFileId);
-            var agreementStatus = Enum.Parse<AgreementStatusTypes>(agreementCurrentStatusTypeCode);
 
-            if (!_statusSolver.CanEditOrDeleteAgreement(currentAcquisitionStatus, agreementStatus) && !_user.HasPermission(Permissions.SystemAdmin))
+            if (!_statusSolver.CanEditOrDeleteAgreement(currentAcquisitionStatus))
             {
-                throw new BusinessRuleViolationException("The file you are editing is not active or hold, so you cannot save changes. Refresh your browser to see file state.");
+                throw new BusinessRuleViolationException("The file you are editing is not active, so you cannot save changes. Refresh your browser to see file state.");
             }
         }
 
@@ -835,36 +859,37 @@ namespace Pims.Api.Services
         {
             var currentAcquisitionFile = _acqFileRepository.GetById(acquisitionFile.Internal_Id);
             var compensationRequisitions = _compensationRequisitionRepository.GetAllByAcquisitionFileId(acquisitionFile.Internal_Id);
+            var compReqAcqPayees = compensationRequisitions.SelectMany(x => x.PimsCompReqAcqPayees).ToList();
 
-            if (compensationRequisitions.Count == 0)
+            if (compReqAcqPayees.Count == 0)
             {
                 return;
             }
 
-            foreach (var compReq in compensationRequisitions)
+            foreach (var payee in compReqAcqPayees)
             {
                 // Check for Acquisition File Owner removed
-                if (compReq.AcquisitionOwnerId is not null
-                    && !acquisitionFile.PimsAcquisitionOwners.Any(x => x.Internal_Id.Equals(compReq.AcquisitionOwnerId))
-                    && currentAcquisitionFile.PimsAcquisitionOwners.Any(x => x.Internal_Id.Equals(compReq.AcquisitionOwnerId)))
+                if (payee.AcquisitionOwnerId is not null
+                    && !acquisitionFile.PimsAcquisitionOwners.Any(x => x.Internal_Id.Equals(payee.AcquisitionOwnerId))
+                    && currentAcquisitionFile.PimsAcquisitionOwners.Any(x => x.Internal_Id.Equals(payee.AcquisitionOwnerId)))
                 {
-                    throw new ForeignKeyDependencyException("Acquisition File Owner can not be removed since it's assigned as a payee for a compensation requisition");
+                    throw new BusinessRuleViolationException("Acquisition File Owner can not be removed since it's assigned as a payee for a compensation requisition");
                 }
 
-                // Check for Acquisition InterestHolders
-                if (compReq.InterestHolderId is not null
-                    && !acquisitionFile.PimsInterestHolders.Any(x => x.Internal_Id.Equals(compReq.InterestHolderId))
-                    && currentAcquisitionFile.PimsInterestHolders.Any(x => x.Internal_Id.Equals(compReq.InterestHolderId)))
+                //// Check for Acquisition InterestHolders
+                if (payee.InterestHolderId is not null
+                    && !acquisitionFile.PimsInterestHolders.Any(x => x.Internal_Id.Equals(payee.InterestHolderId))
+                    && currentAcquisitionFile.PimsInterestHolders.Any(x => x.Internal_Id.Equals(payee.InterestHolderId)))
                 {
-                    throw new ForeignKeyDependencyException("Acquisition File Interest Holders can not be removed since it's assigned as a payee for a compensation requisition");
+                    throw new BusinessRuleViolationException("Acquisition File Interest Holders can not be removed since it's assigned as a payee for a compensation requisition");
                 }
 
-                // Check for File Person
-                if (compReq.AcquisitionFileTeamId is not null
-                    && !acquisitionFile.PimsAcquisitionFileTeams.Any(x => x.Internal_Id.Equals(compReq.AcquisitionFileTeamId))
-                    && currentAcquisitionFile.PimsAcquisitionFileTeams.Any(x => x.Internal_Id.Equals(compReq.AcquisitionFileTeamId)))
+                //// Check for File Person
+                if (payee.AcquisitionFileTeamId is not null
+                    && !acquisitionFile.PimsAcquisitionFileTeams.Any(x => x.Internal_Id.Equals(payee.AcquisitionFileTeamId))
+                    && currentAcquisitionFile.PimsAcquisitionFileTeams.Any(x => x.Internal_Id.Equals(payee.AcquisitionFileTeamId)))
                 {
-                    throw new ForeignKeyDependencyException("Acquisition File team member can not be removed since it's assigned as a payee for a compensation requisition");
+                    throw new BusinessRuleViolationException("Acquisition File team member can not be removed since it's assigned as a payee for a compensation requisition");
                 }
             }
         }
@@ -873,20 +898,21 @@ namespace Pims.Api.Services
         {
             var currentAcquisitionFile = _acqFileRepository.GetById(acquisitionFileId);
             var compensationRequisitions = _compensationRequisitionRepository.GetAllByAcquisitionFileId(acquisitionFileId);
+            var compReqAcqPayees = compensationRequisitions.SelectMany(x => x.PimsCompReqAcqPayees).ToList();
 
-            if (compensationRequisitions.Count == 0)
+            if (compReqAcqPayees.Count == 0)
             {
                 return;
             }
 
-            foreach (var compReq in compensationRequisitions)
+            foreach (var payee in compReqAcqPayees)
             {
                 // Check for Interest Holder
-                if (compReq.InterestHolderId is not null
-                    && !interestHolders.Any(x => x.InterestHolderId.Equals(compReq.InterestHolderId))
-                    && currentAcquisitionFile.PimsInterestHolders.Any(x => x.Internal_Id.Equals(compReq.InterestHolderId)))
+                if (payee.InterestHolderId is not null
+                    && !interestHolders.Any(x => x.InterestHolderId.Equals(payee.InterestHolderId))
+                    && currentAcquisitionFile.PimsInterestHolders.Any(x => x.Internal_Id.Equals(payee.InterestHolderId)))
                 {
-                    throw new ForeignKeyDependencyException("Acquisition File Interest Holder can not be removed since it's assigned as a payee for a compensation requisition");
+                    throw new BusinessRuleViolationException("Acquisition File Interest Holder can not be removed since it's assigned as a payee for a compensation requisition");
                 }
             }
         }
